WIP: kalman_step
