+++ include/filters/kalman.h
@@ -3,28 +3,33 @@
 
 #include "arm_math.h"
 
-typedef struct {
-    arm_matrix_instance_f32 F; // 4 >< 4  const       // State transition matrix
-    arm_matrix_instance_f32 H; // 2 >< 4  const       // Measurement matrix
-    arm_matrix_instance_f32 Q; // 4 >< 4  const       // Process noise covariance
-    arm_matrix_instance_f32 R; // 2 >< 2  const       // Measurement noise covariance
-    arm_matrix_instance_f32 B; // 4 >< 2  const       // Control matrix
-    arm_matrix_instance_f32 I; // 4 >< 4  const Identity
-} KalmanState;
+#define DELTA_T 0.04f
+
+void print_matrix(const arm_matrix_instance_f32 *m, const char *name); 
+
+void print_vector(const float32_t *vec, uint32_t size) ;
+
 
 typedef struct {
-    float32_t vec_X[4];      // x: [pitch, roll, bias_wE, bias_wN]
-    float32_t vec_Z[2];      // x: [measuredPitch, measuredRoll]
-    arm_matrix_instance_f32 P; // 4 >< 4
-    arm_matrix_instance_f32 K; // 4 >< 2        // Kalman gain
-    KalmanState * state;
-} KalmanFilter;
+    // Матрицы, которые не изменяются
+    const arm_matrix_instance_f32 *F;  // Матрица динамики системы
+    const arm_matrix_instance_f32 *B;  // Матрица ypravl
+    const arm_matrix_instance_f32 *Q;  // Матрица noise provess
+    const arm_matrix_instance_f32 *H;  // Матрица nablyde
+    const arm_matrix_instance_f32 *R;  // Матрица ковариации шума измерения
+    const arm_matrix_instance_f32 *I;  // Identity matrix
+    // Матрицы, которые изменяются
+    arm_matrix_instance_f32 *P;  // Covariaton 
+    arm_matrix_instance_f32 *K;  // Kalman gain
+    // Vectors, которые изменяются
+    float32_t *vec_X;
+    float32_t *vec_Z;
+    float32_t *vec_U;
+    
+} KalmanFilter_t;
 
+void kalman_init(KalmanFilter_t *kf, float32_t pitch, float32_t roll); 
 
-int Kalman_Init(KalmanFilter* kf, float32_t initialPitch, float32_t initialRoll);
-// void Kalman_Predict(KalmanFilter* kf, float32_t dt, float32_t wN, float32_t wE);
-// void Kalman_Update(KalmanFilter* kf, float32_t measuredPitch, float32_t measuredRoll);
-// void Kalman_GetAngles(KalmanFilter* kf, float32_t* pitch, float32_t* roll);
-
+void kalman_step(KalmanFilter_t *kf) ;
 #endif
